0.1.3 build - Gold added, Editor Script for scene setup - More...
New monster animations, monsters from stage 1-5 redesigned with more abilities. New special effect (concussion)

+++ Quizzos/Assets/Player/Level System/LevelSystem.cs
@@ -39,7 +39,7 @@ public class LevelSystem : MonoBehaviour
         
     }
 
-    public void GainExp(float expPointsGained)
+    public void GainExp(int expPointsGained)
     {
         ExperiencePoints += expPointsGained;
         if(ExperiencePoints >= xpRequirementsForLevelUp[PlayerLevel - 1])
